refactor: 目標時間設定に関する内容をコンポーネントに分離
